Refactor classic cards presenter into modular views
